Working In Lab 3_02_20
Machine Learning MLOOP Code Updated to operate effecftivley in the lab. Issues regarding writing the voltages were resolved and some data has been taken and stored in MLOOP archives. Data for Adam was also taken although the cooling laser was drifting. A text file for best manual human improvement is saved in the research drive for 3_02_20.